# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API description",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/login": {
            "post": {
                "summary": "Authenticate user",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "summary": "Register a new user",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/expense-categories": {
            "get": {
                "summary": "Get a list of expense categories",
                "operationId": "3a57961c189fa263f2f2ee032139f434",
                "responses": {
                    "200": {
                        "description": "List of expense categories"
                    }
                }
            },
            "post": {
                "summary": "Create a new expense category",
                "operationId": "e9c91d5e2ba41405bebcb5f3781cb86a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense category created"
                    },
                    "400": {
                        "description": "Expense category already exists"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/expense-categories/{id}": {
            "get": {
                "summary": "Get details of a specific expense category",
                "operationId": "0715450f6ef6057c18e5a718092b64be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense category details"
                    },
                    "404": {
                        "description": "Expense category not found"
                    }
                }
            },
            "put": {
                "summary": "Update a specific expense category",
                "operationId": "b63dae36bfd9d298798693cad380466e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense category updated"
                    },
                    "404": {
                        "description": "Expense category not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "summary": "Delete a specific expense category",
                "operationId": "3ede5ddac795968a7e80a1e4494435db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Expense category deleted"
                    },
                    "404": {
                        "description": "Expense category not found"
                    }
                }
            }
        },
...
    }
}